backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
djoser==2.1.0
djangorestframework-simplejwt==4.7.2 
webcolors==1.11.1
Pillow==9.0.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
drf_extra_fields==3.4.0